streamlit_app/requirements.txt
streamlit
requests
pyyaml
openai
azure-identity
azure-keyvault
streamlit-javascript
llama-index
llama-index-llms-azure-openai
llama-index-embeddings-azure-openai
llama-index-readers-azstorage-blob
openpyxl
docx2txt
langchain_community
azure-ai-ml
azure-ai-documentintelligence
azure-storage-blob
PyMuPDF
transformers